+++ src/lib.rs
@@ -2,6 +2,7 @@ pub mod command;
 pub mod dcm_worklist;
 pub mod dcm_xml;
 pub mod gdt;
+pub mod gui;
 pub mod opp_xml;
 pub mod vdds;
 pub mod worklist_conversion;
